trying to fix skybuff issue

+++ smpshift.py
@@ -2783,16 +2783,16 @@ class smp:
                 if badflag:
                     print 'badflaggg'*100
 
-                if not np.min(im[ysn-2:ysn+3,xsn-2:xsn+3]) != np.max(im[ysn-2:ysn+3,xsn-2:xsn+3]):
+                if not np.min(im[int(ysn-2):int(ysn+3),int(xsn-2):int(xsn+3)]) != np.max(im[int(ysn-2):int(ysn+3),int(xsn-2):int(xsn+3)]):
                     print 'im min max'*100
                 if not np.max(psf_stamp[params.substamp / 2 + 1 - 3:params.substamp / 2 + 1 + 4,
                           params.substamp / 2 + 1 - 3:params.substamp / 2 + 1 + 4]) == np.max(psf_stamp[:, :]):
                     print 'psf min max'*100
                 if not badflag:
                     if fwhm_arcsec < params.fwhm_max:
-                        if np.min(im[ysn-2:ysn+3,xsn-2:xsn+3]) != np.max(im[ysn-2:ysn+3,xsn-2:xsn+3]):
+                        if np.min(im[int(ysn-2):int(ysn+3),int(xsn-2):int(xsn+3)]) != np.max(im[int(ysn-2):int(ysn+3),int(xsn-2):int(xsn+3)]):
                             #if len(np.where(mask[ysn-25:ysn+26,xsn-25:xsn+26] != 0)[0]) < params.max_masknum
-                                if np.max(psf_stamp[params.substamp/2+1-3:params.substamp/2+1+4,params.substamp/2+1-3:params.substamp/2+1+4]) == np.max(psf_stamp[:,:]):
+                                if np.max(psf_stamp[int(params.substamp/2+1-3):int(params.substamp/2+1+4),int(params.substamp/2+1-3):int(params.substamp/2+1+4)]) == np.max(psf_stamp[:,:]):
                                     #i = j
